+++ packages/perseus-editor/src/__stories__/flags-for-api-options.ts
@@ -4,7 +4,10 @@ export const flags = {
     mafs: {
         segment: true,
         circle: true,
+        quadratic: true,
+        sinusoid: true,
         "interactive-graph-locked-features-m1": true,
+        "interactive-graph-locked-features-m2": true,
     },
 } satisfies APIOptions["flags"];