+++ packages/perseus-editor/src/components/locked-line-settings.tsx
@@ -4,10 +4,11 @@
  *
  * Used in the interactive graph editor's locked figures section.
  */
+import {vector as kvector} from "@khanacademy/kmath";
 import {View, useUniqueIdWithMock} from "@khanacademy/wonder-blocks-core";
 import {OptionItem, SingleSelect} from "@khanacademy/wonder-blocks-dropdown";
 import {Strut} from "@khanacademy/wonder-blocks-layout";
-import {spacing} from "@khanacademy/wonder-blocks-tokens";
+import {color as wbColor, spacing} from "@khanacademy/wonder-blocks-tokens";
 import {LabelMedium, LabelLarge} from "@khanacademy/wonder-blocks-typography";
 import {StyleSheet} from "aphrodite";
 import * as React from "react";
@@ -25,6 +26,8 @@ import type {
     LockedLineType,
     LockedPointType,
 } from "@khanacademy/perseus";
+
+const lengthZeroStr = "The line cannot have length 0.";
 
 export type Props = LockedLineType &
     AccordionProps & {
@@ -60,6 +63,9 @@ const LockedLineSettings = (props: Props) => {
     const capitalizeKind = kind.charAt(0).toUpperCase() + kind.slice(1);
     const lineLabel = `${capitalizeKind} (${point1.coord[0]},
         ${point1.coord[1]}), (${point2.coord[0]}, ${point2.coord[1]})`;
+
+    // Check if the line has length 0.
+    const isInvalid = kvector.equal(point1.coord, point2.coord);
 
     function handleChangePoint(
         newPointProps: Partial<LockedPointType>,
@@ -123,8 +129,8 @@ const LockedLineSettings = (props: Props) => {
                     placeholder=""
                 >
                     <OptionItem value="line" label="line" />
-                    <OptionItem value="segment" label="segment" />
                     <OptionItem value="ray" label="ray" />
+                    <OptionItem value="segment" label="segment" />
                 </SingleSelect>
             </View>
 
@@ -161,10 +167,18 @@ const LockedLineSettings = (props: Props) => {
                 </View>
             </View>
 
+            {/* Points error message */}
+            {isInvalid && (
+                <LabelMedium style={styles.errorText}>
+                    {lengthZeroStr}
+                </LabelMedium>
+            )}
+
             {/* Defining points settings */}
             <DefiningPointSettings
                 label="Point 1"
                 showPoint={showPoint1}
+                error={isInvalid ? lengthZeroStr : null}
                 {...point1}
                 onTogglePoint={(newValue) =>
                     onChangeProps({showPoint1: newValue})
@@ -174,6 +188,7 @@ const LockedLineSettings = (props: Props) => {
             <DefiningPointSettings
                 label="Point 2"
                 showPoint={showPoint2}
+                error={isInvalid ? lengthZeroStr : null}
                 {...point2}
                 onTogglePoint={(newValue) =>
                     onChangeProps({showPoint2: newValue})
@@ -206,6 +221,9 @@ const styles = StyleSheet.create({
     selectMarginOffset: {
         // Align with the point settings accordions.
         marginInlineEnd: -spacing.xxxSmall_4,
+    },
+    errorText: {
+        color: wbColor.red,
     },
 });