+++ packages/perseus-editor/src/components/util.ts
@@ -76,6 +76,24 @@ export function getDefaultFigureForType(type: LockedFigureType): LockedFigure {
                 showPoint1: false,
                 showPoint2: false,
             };
+        case "circle":
+            return {
+                type: "circle",
+                center: [0, 0],
+                radius: 1,
+                color: DEFAULT_COLOR,
+                fillStyle: "none",
+                strokeStyle: "solid",
+            };
+        case "vector":
+            return {
+                type: "vector",
+                points: [
+                    [0, 0],
+                    [2, 2],
+                ],
+                color: DEFAULT_COLOR,
+            };
         default:
             throw new UnreachableCaseError(type);
     }