+++ packages/perseus/src/perseus-types.ts
@@ -8,11 +8,8 @@ import type {Interval, vec} from "mafs";
 // and exported from the package, so we need to keep it around.
 export type Range = Interval;
 export type Size = [number, number];
-export type CollinearTuple = readonly [vec.Vector2, vec.Vector2];
+export type CollinearTuple = [vec.Vector2, vec.Vector2];
 export type ShowSolutions = "all" | "selected" | "none";
-
-// TODO(FEI-5054): Figure out how to get global .d.ts files working with monorepos
-type Empty = Record<never, never>;
 
 export type PerseusWidgetsMap = {
     [key in `categorizer ${number}`]: CategorizerWidget;
@@ -673,7 +670,11 @@ export const lockedFigureColors: Record<LockedFigureColor, string> = {
     red: "#D92916",
 } as const;
 
-export type LockedFigure = LockedPointType | LockedLineType;
+export type LockedFigure =
+    | LockedPointType
+    | LockedLineType
+    | LockedCircleType
+    | LockedVectorType;
 export type LockedFigureType = LockedFigure["type"];
 
 export type LockedPointType = {
@@ -691,6 +692,21 @@ export type LockedLineType = {
     lineStyle: "solid" | "dashed";
     showPoint1: boolean;
     showPoint2: boolean;
+};
+
+export type LockedCircleType = {
+    type: "circle";
+    center: Coord;
+    radius: number;
+    color: LockedFigureColor;
+    fillStyle: "none" | "solid" | "translucent";
+    strokeStyle: "solid" | "dashed";
+};
+
+export type LockedVectorType = {
+    type: "vector";
+    points: [tail: Coord, tip: Coord];
+    color: LockedFigureColor;
 };
 
 export type PerseusGraphType =
@@ -742,7 +758,7 @@ export type PerseusGraphTypeLinear = {
 export type PerseusGraphTypeLinearSystem = {
     type: "linear-system";
     // expects 2 sets of 2 coords
-    coords?: ReadonlyArray<CollinearTuple>;
+    coords?: CollinearTuple[];
 } & PerseusGraphTypeCommon;
 
 export type PerseusGraphTypePoint = {
@@ -770,7 +786,7 @@ export type PerseusGraphTypePolygon = {
 export type PerseusGraphTypeQuadratic = {
     type: "quadratic";
     // expects a list of 3 coords
-    coords?: ReadonlyArray<Coord>;
+    coords?: [Coord, Coord, Coord];
 } & PerseusGraphTypeCommon;
 
 export type PerseusGraphTypeSegment = {
@@ -778,7 +794,7 @@ export type PerseusGraphTypeSegment = {
     // The number of segments if a "segment" type. default: 1.  Max: 6
     numSegments?: number;
     // Expects a list of Coord tuples. Length should match the `numSegments` value.
-    coords?: ReadonlyArray<CollinearTuple>;
+    coords?: CollinearTuple[];
 } & PerseusGraphTypeCommon;
 
 export type PerseusGraphTypeSinusoid = {