+++ packages/perseus/src/widgets/__tests__/interactive-graph.test.ts
@@ -8,6 +8,7 @@ import {testDependencies} from "../../../../../testing/test-dependencies";
 import * as Dependencies from "../../dependencies";
 import {ApiOptions} from "../../perseus-api";
 import {lockedFigureColors} from "../../perseus-types";
+import {sinusoidQuestion} from "../__testdata__/grapher.testdata";
 import {
     circleQuestion,
     circleQuestionWithDefaultCorrect,
@@ -19,14 +20,19 @@ import {
     pointQuestionWithDefaultCorrect,
     polygonQuestion,
     polygonQuestionDefaultCorrect,
+    quadraticQuestion,
+    quadraticQuestionWithDefaultCorrect,
     questionsAndAnswers,
     rayQuestion,
     rayQuestionWithDefaultCorrect,
     segmentQuestion,
     segmentQuestionDefaultCorrect,
+    segmentWithLockedCircles,
     segmentWithLockedLineQuestion,
     segmentWithLockedPointsQuestion,
     segmentWithLockedPointsWithColorQuestion,
+    segmentWithLockedVectors,
+    sinusoidQuestionWithDefaultCorrect,
 } from "../__testdata__/interactive-graph.testdata";
 import {trueForAllMafsSupportedGraphTypes} from "../interactive-graphs/mafs-supported-graph-types";
 
@@ -154,6 +160,8 @@ describe("a mafs graph", () => {
         polygon: polygonQuestion,
         point: pointQuestion,
         circle: circleQuestion,
+        quadratic: quadraticQuestion,
+        sinusoid: sinusoidQuestion,
     };
 
     const graphQuestionRenderersCorrect: {
@@ -166,6 +174,8 @@ describe("a mafs graph", () => {
         polygon: polygonQuestionDefaultCorrect,
         point: pointQuestionWithDefaultCorrect,
         circle: circleQuestionWithDefaultCorrect,
+        quadratic: quadraticQuestionWithDefaultCorrect,
+        sinusoid: sinusoidQuestionWithDefaultCorrect,
     };
 
     describe.each(Object.entries(graphQuestionRenderers))(
@@ -527,11 +537,10 @@ describe("locked layer", () => {
             stroke: lockedFigureColors.pink,
         });
     });
-});
 
-describe("snapshots", () => {
-    test("should render correctly", () => {
-        const {container} = renderQuestion(segmentQuestionDefaultCorrect, {
+    test("should render locked circles", async () => {
+        // Arrange
+        const {container} = renderQuestion(segmentWithLockedCircles, {
             flags: {
                 mafs: {
                     segment: true,
@@ -539,11 +548,29 @@ describe("snapshots", () => {
             },
         });
 
-        expect(container).toMatchSnapshot();
+        // Act
+        // eslint-disable-next-line testing-library/no-container, testing-library/no-node-access
+        const circles = container.querySelectorAll("ellipse");
+
+        // Assert
+        expect(circles).toHaveLength(3);
+        expect(circles[0]).toHaveStyle({
+            "fill-opacity": "0",
+            stroke: lockedFigureColors["grayH"],
+        });
+        expect(circles[1]).toHaveStyle({
+            "fill-opacity": "1",
+            stroke: lockedFigureColors["green"],
+        });
+        expect(circles[2]).toHaveStyle({
+            "fill-opacity": "0.4",
+            stroke: lockedFigureColors["green"],
+        });
     });
 
-    test("should render correctly with locked points", () => {
-        const {container} = renderQuestion(segmentWithLockedPointsQuestion, {
+    test("should render locked vectors", async () => {
+        // Arrange
+        const {container} = renderQuestion(segmentWithLockedVectors, {
             flags: {
                 mafs: {
                     segment: true,
@@ -551,18 +578,41 @@ describe("snapshots", () => {
             },
         });
 
-        expect(container).toMatchSnapshot();
-    });
+        // Act
+        // eslint-disable-next-line testing-library/no-container, testing-library/no-node-access
+        const vectors = container.querySelectorAll(".locked-vector");
 
-    test("should render correctly with locked lines", () => {
-        const {container} = renderQuestion(segmentWithLockedLineQuestion, {
-            flags: {
-                mafs: {
-                    segment: true,
-                },
-            },
+        // Assert
+        expect(vectors).toHaveLength(2);
+        // eslint-disable-next-line testing-library/no-container, testing-library/no-node-access
+        let vector = vectors[0].children[0];
+        expect(vector).toHaveStyle({
+            "stroke-width": "2",
+            stroke: lockedFigureColors["grayH"],
         });
+        // eslint-disable-next-line testing-library/no-container, testing-library/no-node-access
+        let arrowheads = vector.querySelectorAll(
+            ".interactive-graph-arrowhead",
+        );
+        expect(arrowheads).toHaveLength(1);
+        // Arrowhead should be at the end (tip) of the vector, and rotated
+        expect(arrowheads[0]).toHaveAttribute(
+            "transform",
+            "translate(40 -40) rotate(-45)",
+        );
 
-        expect(container).toMatchSnapshot();
+        // eslint-disable-next-line testing-library/no-container, testing-library/no-node-access
+        vector = vectors[1].children[0];
+        expect(vector).toHaveStyle({
+            "stroke-width": "2",
+            stroke: lockedFigureColors["green"],
+        });
+        // eslint-disable-next-line testing-library/no-container, testing-library/no-node-access
+        arrowheads = vector.querySelectorAll(".interactive-graph-arrowhead");
+        expect(arrowheads).toHaveLength(1);
+        expect(arrowheads[0]).toHaveAttribute(
+            "transform",
+            "translate(-40 -80) rotate(-153.43494882292202)",
+        );
     });
 });