+++ packages/perseus/src/widgets/interactive-graphs/graph-locked-layer.tsx
@@ -1,8 +1,10 @@
 import {UnreachableCaseError} from "@khanacademy/wonder-stuff-core";
 import * as React from "react";
 
+import LockedCircle from "./locked-circle";
 import LockedLine from "./locked-line";
 import LockedPoint from "./locked-point";
+import LockedVector from "./locked-vector";
 
 import type {LockedFigure} from "../../perseus-types";
 import type {Interval} from "mafs";
@@ -29,6 +31,14 @@ const GraphLockedLayer = (props: Props) => {
                                 range={props.range}
                                 {...figure}
                             />
+                        );
+                    case "circle":
+                        return (
+                            <LockedCircle key={`circle-${index}`} {...figure} />
+                        );
+                    case "vector":
+                        return (
+                            <LockedVector key={`vector-${index}`} {...figure} />
                         );
                     default:
                         /**