+++ packages/perseus/src/widgets/interactive-graphs/graphs/components/arrowhead.tsx
@@ -27,7 +27,7 @@ export function Arrowhead(props: Props) {
     return (
         <g
             className="interactive-graph-arrowhead"
-            transform={`translate(${point[0]} ${point[1]}) rotate(${-props.angle})`}
+            transform={`translate(${point[0]} ${point[1]}) rotate(${props.angle})`}
         >
             <g transform="translate(-1.5)">
                 <path
@@ -36,6 +36,7 @@ export function Arrowhead(props: Props) {
                     style={{stroke: props.color ?? "inherit"}}
                     strokeLinejoin="round"
                     strokeLinecap="round"
+                    strokeWidth="2px"
                 />
             </g>
         </g>