+++ packages/perseus/src/widgets/interactive-graphs/reducer/interactive-graph-reducer.ts
@@ -3,6 +3,7 @@ import {UnreachableCaseError} from "@khanacademy/wonder-stuff-core";
 import {vec} from "mafs";
 import _ from "underscore";
 
+import {polygonSidesIntersect} from "../../../util/geometry";
 import {snap} from "../utils";
 
 import {
@@ -58,7 +59,6 @@ function doMoveControlPoint(
     state: InteractiveGraphState,
     action: MoveControlPoint,
 ): InteractiveGraphState {
-    const {snapStep, range} = state;
     switch (state.type) {
         case "segment":
         case "linear":
@@ -71,14 +71,7 @@ function doMoveControlPoint(
                     setAtIndex({
                         array: tuple,
                         index: action.pointIndex,
-                        newValue: snap(
-                            snapStep,
-                            bound({
-                                snapStep,
-                                range,
-                                point: action.destination,
-                            }),
-                        ),
+                        newValue: boundAndSnapToGrid(action.destination, state),
                     }),
             });
 
@@ -96,6 +89,8 @@ function doMoveControlPoint(
             throw new Error("FIXME implement circle reducer");
         case "point":
         case "polygon":
+        case "quadratic":
+        case "sinusoid":
             throw new Error(
                 `Don't use moveControlPoint for ${state.type} graphs. Use movePoint instead!`,
             );
@@ -184,28 +179,67 @@ function doMovePoint(
     action: MovePoint,
 ): InteractiveGraphState {
     switch (state.type) {
-        case "point":
-        case "polygon": {
+        case "polygon":
+            const newCoords = setAtIndex({
+                array: state.coords,
+                index: action.index,
+                newValue: boundAndSnapToGrid(action.destination, state),
+            });
+
+            // Reject the move if it would cause the sides of the polygon to cross
+            if (polygonSidesIntersect(newCoords)) {
+                return state;
+            }
+
+            return {
+                ...state,
+                hasBeenInteractedWith: true,
+                coords: newCoords,
+            };
+        case "point": {
             return {
                 ...state,
                 hasBeenInteractedWith: true,
                 coords: setAtIndex({
                     array: state.coords,
                     index: action.index,
-                    newValue: snap(
-                        state.snapStep,
-                        bound({
-                            snapStep: state.snapStep,
-                            range: state.range,
-                            point: action.destination,
-                        }),
-                    ),
+                    newValue: boundAndSnapToGrid(action.destination, state),
+                }),
+            };
+        }
+        case "sinusoid": {
+            // First, we need to verify that the new coordinates are not on the same vertical line
+            // If they are, we don't want to move the point
+            const destination = action.destination;
+            const newCoords: vec.Vector2[] = [...state.coords];
+            newCoords[action.index] = action.destination;
+            if (newCoords[0][0] === newCoords[1][0]) {
+                return state;
+            }
+            return {
+                ...state,
+                hasBeenInteractedWith: true,
+                coords: setAtIndex({
+                    array: state.coords,
+                    index: action.index,
+                    newValue: boundAndSnapToGrid(destination, state),
+                }),
+            };
+        }
+        case "quadratic": {
+            return {
+                ...state,
+                hasBeenInteractedWith: true,
+                coords: setAtIndex({
+                    array: state.coords,
+                    index: action.index,
+                    newValue: boundAndSnapToGrid(action.destination, state),
                 }),
             };
         }
         default:
             throw new Error(
-                "The movePoint action is only for point and polygon graphs",
+                "The movePoint action is only for point, quadratic, and polygon graphs",
             );
     }
 }
@@ -350,6 +384,13 @@ interface ConstraintArgs {
     snapStep: vec.Vector2;
     range: [Interval, Interval];
     point: vec.Vector2;
+}
+
+function boundAndSnapToGrid(
+    point: vec.Vector2,
+    {snapStep, range}: {snapStep: vec.Vector2; range: [Interval, Interval]},
+) {
+    return snap(snapStep, bound({snapStep, range, point}));
 }
 
 // Returns the closest point to the given `point` that is within the graph