+++ packages/perseus/src/widgets/interactive-graphs/reducer/interactive-graph-state.ts
@@ -1,4 +1,5 @@
 import {UnreachableCaseError} from "@khanacademy/wonder-stuff-core";
+import {vec} from "mafs";
 
 import {normalizeCoords, normalizePoints} from "../utils";
 
@@ -10,6 +11,9 @@ import type {
     PerseusGraphTypeLinear,
     PerseusGraphTypeLinearSystem,
     PerseusGraphTypePolygon,
+    PerseusGraphTypeQuadratic,
+    PerseusGraphTypeSinusoid,
+    PerseusGraphTypeCircle,
 } from "../../../perseus-types";
 import type {
     CircleGraphState,
@@ -24,7 +28,7 @@ export type InitializeGraphStateParam = {
     range: InteractiveGraphProps["range"];
     step: InteractiveGraphProps["step"];
     snapStep: InteractiveGraphProps["snapStep"];
-    graph: InteractiveGraphProps["graph"];
+    graph: PerseusGraphType;
 };
 
 export function initializeGraphState(
@@ -41,17 +45,25 @@ export function initializeGraphState(
             return {
                 ...shared,
                 type: "segment",
-                coords: getDefaultSegments({graph, step, range}),
+                coords: getSegmentCoords({graph, step, range}),
             };
         case "linear":
-        case "linear-system":
+            return {
+                ...shared,
+                type: graph.type,
+                coords: getLineCoords({graph, range, step}),
+            };
         case "ray":
             return {
                 ...shared,
                 type: graph.type,
-                // Linear and ray graphs have a single tuple of points, while a
-                // linear system has two tuples of points.
                 coords: getLineCoords({graph, range, step}),
+            };
+        case "linear-system":
+            return {
+                ...shared,
+                type: graph.type,
+                coords: getLinearSystemCoords({graph, range, step}),
             };
         case "polygon":
             return {
@@ -71,12 +83,21 @@ export function initializeGraphState(
             return {
                 ...shared,
                 type: graph.type,
-                center: [0, 0],
-                radiusPoint: [1, 0],
+                ...getCircleCoords(graph, range, step),
+            };
+        case "quadratic":
+            return {
+                ...shared,
+                type: graph.type,
+                coords: getQuadraticCoords(graph, range, step),
+            };
+        case "sinusoid":
+            return {
+                ...shared,
+                type: graph.type,
+                coords: getSinusoidCoords(graph, range, step),
             };
         case "angle":
-        case "sinusoid":
-        case "quadratic":
             throw new Error(
                 "Mafs not yet implemented for graph type: " + graph.type,
             );
@@ -222,6 +243,20 @@ export function getGradableGraph(
         };
     }
 
+    if (state.type === "quadratic" && initialGraph.type === "quadratic") {
+        return {
+            ...initialGraph,
+            coords: state.coords,
+        };
+    }
+
+    if (state.type === "sinusoid" && initialGraph.type === "sinusoid") {
+        return {
+            ...initialGraph,
+            coords: state.coords,
+        };
+    }
+
     throw new Error(
         "Mafs is not yet implemented for graph type: " + initialGraph.type,
     );
@@ -233,11 +268,15 @@ type getDefaultSegmentsArg = {
     step: InitializeGraphStateParam["step"];
 };
 
-const getDefaultSegments = ({
+const getSegmentCoords = ({
     graph,
     range,
     step,
 }: getDefaultSegmentsArg): PairOfPoints[] => {
+    if (graph.coords) {
+        return graph.coords;
+    }
+
     const ys = (n?: number) => {
         switch (n) {
             case 2:
@@ -283,10 +322,7 @@ const defaultLinearCoords: [Coord, Coord][] = [
 ];
 
 type getLineCoordsArg = {
-    graph:
-        | PerseusGraphTypeRay
-        | PerseusGraphTypeLinear
-        | PerseusGraphTypeLinearSystem;
+    graph: PerseusGraphTypeRay | PerseusGraphTypeLinear;
     range: InitializeGraphStateParam["range"];
     step: InitializeGraphStateParam["step"];
 };
@@ -295,13 +331,33 @@ const getLineCoords = ({
     graph,
     range,
     step,
-}: getLineCoordsArg): PairOfPoints[] =>
-    // Return two lines for a linear system, one for a ray or linear
-    graph.coords ?? graph.type === "linear-system"
-        ? defaultLinearCoords.map((collinear) =>
-              normalizePoints(range, step, collinear),
-          )
-        : [normalizePoints(range, step, defaultLinearCoords[0])];
+}: getLineCoordsArg): PairOfPoints[] => {
+    if (graph.coords) {
+        return [graph.coords];
+    }
+
+    return [normalizePoints(range, step, defaultLinearCoords[0])];
+};
+
+type getLinearSystemCoordsArg = {
+    graph: PerseusGraphTypeLinearSystem;
+    range: InitializeGraphStateParam["range"];
+    step: InitializeGraphStateParam["step"];
+};
+
+const getLinearSystemCoords = ({
+    graph,
+    range,
+    step,
+}: getLinearSystemCoordsArg): PairOfPoints[] => {
+    if (graph.coords) {
+        return graph.coords;
+    }
+
+    return defaultLinearCoords.map((points) =>
+        normalizePoints(range, step, points),
+    );
+};
 
 type getPolygonCoordsArg = {
     graph: PerseusGraphTypePolygon;
@@ -347,6 +403,60 @@ const getPolygonCoords = ({
     coords = normalizePoints(range, step, coords, /* noSnap */ !snapToGrid);
 
     return coords;
+};
+
+const getSinusoidCoords = (
+    graph: PerseusGraphTypeSinusoid,
+    range: InitializeGraphStateParam["range"],
+    step: InitializeGraphStateParam["step"],
+): [Coord, Coord] => {
+    if (graph.coords) {
+        return [graph.coords[0], graph.coords[1]];
+    }
+
+    let coords: [Coord, Coord] = [
+        [0.5, 0.5],
+        [0.65, 0.6],
+    ];
+
+    coords = normalizePoints(range, step, coords, true);
+
+    return coords;
+};
+
+const getQuadraticCoords = (
+    graph: PerseusGraphTypeQuadratic,
+    range: InitializeGraphStateParam["range"],
+    step: InitializeGraphStateParam["step"],
+): [Coord, Coord, Coord] => {
+    if (graph.coords) {
+        return graph.coords;
+    }
+
+    const defaultCoords: [Coord, Coord, Coord] = [
+        [0.25, 0.75],
+        [0.5, 0.25],
+        [0.75, 0.75],
+    ];
+
+    return normalizePoints(range, step, defaultCoords, true);
+};
+
+const getCircleCoords = (
+    graph: PerseusGraphTypeCircle,
+    range: InitializeGraphStateParam["range"],
+    step: InitializeGraphStateParam["step"],
+): {center: Coord; radiusPoint: Coord} => {
+    if (graph.center != null && graph.radius != null) {
+        return {
+            center: graph.center,
+            radiusPoint: vec.add(graph.center, [graph.radius, 0]),
+        };
+    }
+    return {
+        center: [0, 0],
+        radiusPoint: [1, 0],
+    };
 };
 
 /**