+++ packages/perseus/src/widgets/interactive-graphs/types.ts
@@ -1,8 +1,5 @@
 import type {InteractiveGraphAction} from "./reducer/interactive-graph-action";
-import type {
-    PerseusGraphType,
-    PerseusInteractiveGraphWidgetOptions,
-} from "../../perseus-types";
+import type {PerseusInteractiveGraphWidgetOptions} from "../../perseus-types";
 import type {WidgetProps} from "../../types";
 import type {Coord} from "@khanacademy/perseus";
 import type {Interval, vec} from "mafs";
@@ -17,19 +14,15 @@ export type MafsGraphProps<T extends InteractiveGraphState> = {
     dispatch: (action: InteractiveGraphAction) => unknown;
 };
 
-export interface InitializeGraphStateParams<T extends PerseusGraphType> {
-    graph: T;
-    range: [Interval, Interval];
-    step: vec.Vector2;
-}
-
 export type InteractiveGraphState =
     | SegmentGraphState
     | LinearGraphState
     | RayGraphState
     | PolygonGraphState
     | PointGraphState
-    | CircleGraphState;
+    | CircleGraphState
+    | QuadraticGraphState
+    | SinusoidGraphState;
 
 export interface InteractiveGraphStateCommon {
     hasBeenInteractedWith: boolean;
@@ -70,6 +63,16 @@ export interface CircleGraphState extends InteractiveGraphStateCommon {
     type: "circle";
     center: Coord;
     radiusPoint: Coord;
+}
+
+export interface QuadraticGraphState extends InteractiveGraphStateCommon {
+    type: "quadratic";
+    coords: [Coord, Coord, Coord];
+}
+
+export interface SinusoidGraphState extends InteractiveGraphStateCommon {
+    type: "sinusoid";
+    coords: [vec.Vector2, vec.Vector2];
 }
 
 export type PairOfPoints = [Coord, Coord];